Build can process nested keys in addons
Almost all plugins require some configuration data to be specified
in the game manifest. This commit adds support for reading from
the "addons" section of the game manifest at build. Keys in this
section may be nested. Specifying a key of the form "foo.bar.baz"
in a plugin config.json will yield the appropriate value.