fix (native-ios): fix builds without ios key
iOS builds fail if there is no ios key in the manifest because the code
is looking for subkeys of the ios key without checking first if the ios
key exists (app.manifest.ios.icons, app.manifest.ios.version).

This fix adds checks for the ios key before checking fields futher down
the hierarchy, which allows the builds to continue without the ios key
in the manifest.