Add new plugin header/lib search path support
The folders `lib` and `include` are now added as a library search path
and header search path, respectively, to the XCode project. The paths
are made relative from the project folder to the plugin folders.
Additionally, the `libs` keyword in `config.json` can be used to specify
a list of static libraries to link against. This is now preferred over
putting them in the `frameworks` list which causes the libraries to be
copied into the project.

The node-xcode project was forked to make this possible. A github
tarball link at the appropriate ref is now listed in the package.json
file instead of the npm package.