Deployment Target iOS 5.1.1 --> 6
iOS 6 is < 5% of users according to many metrics. Also worth noting, the
Facebook SDK only supports iOS7+ and Xcode 7 with the iOS 9 SDK supports
deployment only back to 6.0.  6.0 supports devices back to the iPhone 3GS and
3rd generation iPad.  5.1.1 also only supported back to the iPhone 3GS, so
this only drops support for the first and second generation iPads (as far as I
can tell!).